// src/model/ingrediaents.ts
export type TIngredientTypes = {
    "_id"?: string;
    "name": string;
    "type": string;
    "proteins": number | null;
    "fat": number | null;
    "carbohydrates": number | null;
    "calories": number | null;
    "price": number | null;
    "image"?: string;
    "image_mobile"?: string;
    "image_large"?: string;
    "__v"?: number;
    "count"?: number;
    id?: string;
}